+++ scripts/gen-testdata.sh
@@ -44,6 +44,12 @@ buf build --as-file-descriptor-set --exclude-source-info \
 echo "wrote crates/protowire-pxf/testdata/hardening-test.binpb (buf)"
 
 buf build --as-file-descriptor-set --exclude-source-info \
+  --path crates/protowire-pxf/testdata/schema-test.proto \
+  -o crates/protowire-pxf/testdata/schema-test.binpb
+
+echo "wrote crates/protowire-pxf/testdata/schema-test.binpb (buf)"
+
+buf build --as-file-descriptor-set --exclude-source-info \
   --path crates/protowire-sbe/testdata/sbe-test.proto \
   -o crates/protowire-sbe/testdata/sbe-test.binpb